cli/requirements.txt
# MorningPal CLI Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
